refactor: move target/provider to server
