+++ pkg/cmd/server/provider/list.go
@@ -6,7 +6,7 @@ package provider
 import (
 	"github.com/daytonaio/daytona/internal/util/apiclient/server"
 	"github.com/daytonaio/daytona/pkg/cmd/output"
-	"github.com/daytonaio/daytona/pkg/views/provider"
+	"github.com/daytonaio/daytona/pkg/views/server/provider"
 	"github.com/spf13/cobra"
 
 	log "github.com/sirupsen/logrus"
