+++ pkg/server/api/controllers/provider/install.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 	"path/filepath"
 
-	"github.com/daytonaio/daytona/pkg/provider/manager"
+	"github.com/daytonaio/daytona/pkg/server/provider/manager"
 	"github.com/daytonaio/daytona/pkg/server/api/controllers/provider/dto"
 	"github.com/daytonaio/daytona/pkg/server/api/util"
 	"github.com/daytonaio/daytona/pkg/server/config"
