+++ pkg/server/provisioner/provisioner.go
@@ -10,8 +10,8 @@ import (
 	"github.com/daytonaio/daytona/internal"
 	"github.com/daytonaio/daytona/internal/util"
 	"github.com/daytonaio/daytona/pkg/logger"
-	provider_types "github.com/daytonaio/daytona/pkg/provider"
-	"github.com/daytonaio/daytona/pkg/provider/manager"
+	provider_types "github.com/daytonaio/daytona/pkg/server/provider"
+	"github.com/daytonaio/daytona/pkg/server/provider/manager"
 	"github.com/daytonaio/daytona/pkg/server/config"
 	"github.com/daytonaio/daytona/pkg/server/event_bus"
 	"github.com/daytonaio/daytona/pkg/server/frpc"
